Update file
My new version allows for a user to specify a function to use to determine the distances as well as combine the alpha and beta parameters by multiplying the density and  nearest higher distance metrics. (Note if this is passed than the pydpc.py file will have to be updated for the reduction in parameters.